pyproject.toml
[project]
name = "face-recognize-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Dungpv04", email = "vandungpham2004@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "faiss-cpu>=1.10.0",
    "fastapi[all,standard]>=0.115.11",
    "keras-facenet>=0.3.2",
    "mtcnn>=1.0.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pgvector>=0.3.6",
    "pip>=25.0.1",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "spicy>=0.16.0",
    "sqlalchemy>=2.0.38",
    "sqlmodel>=0.0.23",
    "tensorflow>=2.18.0",
    "uvicorn>=0.34.0",
]

[project.scripts]
dev = "face_recognize_api:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/face_recognize_api"]
